requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.25.0

# Qwen VL dependencies
qwen-vl-utils>=0.0.1

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0

# OCR
rapidocr-onnxruntime>=1.3.0
onnxruntime>=1.16.0

# Audio transcription
openai-whisper>=20231117
ffmpeg-python>=0.2.0

# Utilities
numpy>=1.24.0
